# swagger-api-documentation.yaml
openapi: 3.0.3
info:
  title: Agentic Itinerary Planner API
  description: |
    Comprehensive API for AI-powered travel itinerary planning and management.
    
    This API provides endpoints for:
    - Creating and managing travel itineraries
    - Real-time agent progress tracking via Server-Sent Events
    - Travel planning tools (weather, packing lists, photo spots, etc.)
    - Booking and payment processing
    - Health monitoring and testing
    
    All endpoints support CORS and work without authentication in development mode.
  version: 1.0.0
  contact:
    name: Agentic Itinerary Planner Team
    email: support@agenticitinerary.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://api.agenticitinerary.com/api/v1
    description: Production server
...
paths:
...
  /tools/weather:
    get:
      tags:
        - Tools
      summary: Get weather information
      description: Retrieves weather information for a specific destination
      parameters:
        - name: destination
          in: query
          required: true
          description: Destination name or location
          schema:
            type: string
            example: "Barcelona"
        - name: date
          in: query
          required: false
          description: Date for weather forecast (ISO format)
          schema:
            type: string
            format: date
            example: "2025-06-01"
      responses:
        '200':
          description: Weather information retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeatherRes'
              example:
                destination: "Barcelona"
                date: "2025-06-01"
                temperature: 22.5
                condition: "Sunny"
                description: "Clear skies with light winds"
                humidity: 65.0
                windSpeed: 12.5
                unit: "Celsius"
  /packing-list:
    post:
      tags:
        - Tools
      summary: Generate packing list
      description: Generates a personalized packing list for a trip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PackingListReq'
      responses:
        '200':
          description: Packing list generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackingListRes'
  /photo-spots:
    post:
      tags:
        - Tools
      summary: Get photo spots
      description: Retrieves recommended photo spots for a destination
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PhotoSpotsReq'
      responses:
        '200':
          description: Photo spots retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhotoSpotsRes'
  /must-try-foods:
    post:
      tags:
        - Tools
      summary: Get must-try foods
      description: Retrieves must-try food recommendations for a destination
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MustTryFoodsReq'
      responses:
        '200':
          description: Must-try foods retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MustTryFoodsRes'
  /cost-estimator:
    post:
      tags:
        - Tools
      summary: Generate cost estimate
      description: Generates a cost estimate for a trip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CostEstimateReq'
      responses:
        '200':
          description: Cost estimate generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CostEstimateRes'

  # Booking Endpoints
...
components:
  schemas:
...
    WeatherRes:
      type: object
      properties:
        destination:
          type: string
          description: Destination name
          example: "Barcelona"
        date:
          type: string
          description: Date for weather
          example: "2025-06-01"
        temperature:
          type: number
          format: double
          description: Temperature
          example: 22.5
        condition:
          type: string
          description: Weather condition
          example: "Sunny"
        description:
          type: string
          description: Weather description
          example: "Clear skies with light winds"
        humidity:
          type: number
          format: double
          description: Humidity percentage
          example: 65.0
        windSpeed:
          type: number
          format: double
          description: Wind speed
          example: 12.5
        unit:
          type: string
          description: Temperature unit
          example: "Celsius"
    PackingListReq:
      type: object
      required:
        - destination
        - climate
        - season
        - startDate
        - endDate
        - partySize
      properties:
        destination:
          type: string
          description: Travel destination
          example: "Barcelona"
        climate:
          type: string
          description: Climate type
          example: "Mediterranean"
        season:
          type: string
          description: Travel season
          example: "summer"
        startDate:
          type: string
          format: date
          description: Trip start date
          example: "2025-06-01"
        endDate:
          type: string
          format: date
          description: Trip end date
          example: "2025-06-04"
        activities:
          type: array
          items:
            type: string
          maxItems: 20
          description: Planned activities
          example: ["hiking", "swimming", "sightseeing"]
        partySize:
          type: integer
          minimum: 1
          description: Number of travelers
          example: 3
    PackingListRes:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PackingItem'
    PackingItem:
      type: object
      properties:
        name:
          type: string
          description: Item name
          example: "Sunscreen"
        quantity:
          type: integer
          description: Recommended quantity
          example: 2
        group:
          type: string
          description: Item category
          example: "Health & Safety"
        essential:
          type: boolean
          description: Whether item is essential
          example: true
        notes:
          type: string
          description: Additional notes
          example: "SPF 30 or higher"
    PhotoSpotsReq:
      type: object
      required:
        - destination
      properties:
        destination:
          type: string
          description: Travel destination
          example: "Barcelona"
        interests:
          type: array
          items:
            type: string
          maxItems: 10
          description: Photography interests
          example: ["architecture", "nature", "street"]
        timeOfDay:
          type: string
          description: Preferred time of day
          example: "golden hour"
        season:
          type: string
          description: Travel season
          example: "summer"
    PhotoSpotsRes:
      type: object
      properties:
        spots:
          type: array
          items:
            $ref: '#/components/schemas/PhotoSpot'
    PhotoSpot:
      type: object
      properties:
        name:
          type: string
          description: Photo spot name
          example: "Park Güell"
        lat:
          type: number
          format: double
          description: Latitude
          example: 41.4145
        lng:
          type: number
          format: double
          description: Longitude
          example: 2.1527
        category:
          type: string
          description: Photo category
          example: "architecture"
        bestTime:
          type: string
          description: Best time to visit
          example: "early morning"
        tips:
          type: string
          description: Photography tips
          example: "Bring wide-angle lens"
        difficulty:
          type: string
          description: Access difficulty
          example: "easy"
    MustTryFoodsReq:
      type: object
      required:
        - destination
      properties:
        destination:
          type: string
          description: Travel destination
          example: "Barcelona"
        dietaryPreferences:
          type: array
          items:
            type: string
          maxItems: 10
          description: Dietary preferences
          example: ["vegetarian", "gluten-free"]
        budgetTier:
          type: string
          description: Budget tier
          example: "mid-range"
        cuisineTypes:
          type: array
          items:
            type: string
          description: Preferred cuisine types
          example: ["Spanish", "Catalan"]
    MustTryFoodsRes:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/MustTryFood'
    MustTryFood:
      type: object
      properties:
        name:
          type: string
          description: Food name
          example: "Paella Valenciana"
        description:
          type: string
          description: Food description
          example: "Traditional rice dish with seafood"
        category:
          type: string
          description: Food category
          example: "main course"
        venues:
          type: array
          items:
            type: string
          description: Recommended venues
          example: ["Restaurante 7 Portes", "Can Solé"]
        priceRange:
          type: string
          description: Price range
          example: "€15-25"
        tips:
          type: string
          description: Tips for trying this food
          example: "Best enjoyed for lunch"
    CostEstimateReq:
      type: object
      required:
        - destination
        - startDate
        - endDate
        - budgetTier
        - partySize
      properties:
        destination:
          type: string
          description: Travel destination
          example: "Barcelona"
        startDate:
          type: string
          format: date
          description: Trip start date
          example: "2025-06-01"
        endDate:
          type: string
          format: date
          description: Trip end date
          example: "2025-06-04"
        budgetTier:
          type: string
          description: Budget tier
          example: "mid-range"
        partySize:
          type: integer
          minimum: 1
          description: Number of travelers
          example: 3
        interests:
          type: array
          items:
            type: string
          maxItems: 20
          description: Travel interests
          example: ["culture", "food", "sightseeing"]
    CostEstimateRes:
      type: object
      properties:
        currency:
          type: string
          description: Currency code
          example: "EUR"
        totals:
          $ref: '#/components/schemas/CostBreakdown'
        perDay:
          $ref: '#/components/schemas/CostBreakdown'
        perPerson:
          $ref: '#/components/schemas/CostBreakdown'
        confidence:
          type: string
          description: Estimate confidence level
          example: "high"
        notes:
          type: string
          description: Additional notes
          example: "Prices may vary by season"
    CostBreakdown:
      type: object
      properties:
        transport:
          type: number
          format: double
          description: Transportation costs
          example: 450.0
        lodging:
          type: number
          format: double
          description: Accommodation costs
          example: 600.0
        food:
          type: number
          format: double
          description: Food and dining costs
          example: 300.0
        activities:
          type: number
          format: double
          description: Activity costs
          example: 200.0
        shopping:
          type: number
          format: double
          description: Shopping costs
          example: 150.0
        misc:
          type: number
          format: double
          description: Miscellaneous costs
          example: 100.0
        total:
          type: number
          format: double
          description: Total cost
          example: 1800.0

    # Booking DTOs
...
  securitySchemes:
    # No authentication required in development mode
    # This can be extended for production use
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication (not currently required)
security:
  # No security requirements in development mode
  - {}